+++ lightyear/src/client/mod.rs
@@ -27,4 +27,5 @@ mod easings;
 #[cfg(feature = "leafwing")]
 pub mod input_leafwing;
 pub(crate) mod message;
-pub(crate) mod systems;
+pub(crate) mod networking;
+mod replication;
