+++ lightyear/src/protocol/component.rs
@@ -13,7 +13,7 @@ use serde::{Deserialize, Serialize};
 use crate::client::components::{ComponentSyncMode, LerpFn, SyncMetadata};
 use crate::prelude::{LightyearMapEntities, Message, Named, PreSpawnedPlayerObject};
 use crate::protocol::{BitSerializable, EventContext, Protocol};
-use crate::shared::events::{
+use crate::shared::events::connection::{
     IterComponentInsertEvent, IterComponentRemoveEvent, IterComponentUpdateEvent,
 };
 use crate::shared::replication::components::ShouldBeInterpolated;
