+++ lightyear/src/server/mod.rs
@@ -22,4 +22,6 @@ pub mod room;
 pub mod input_leafwing;
 pub(crate) mod message;
 pub(crate) mod prediction;
-mod systems;
+
+mod networking;
+mod replication;
