+++ lightyear/src/server/room.rs
@@ -12,7 +12,6 @@ use bevy::utils::{HashMap, HashSet};
 use crate::connection::netcode::ClientId;
 use crate::prelude::ReplicationSet;
 use crate::protocol::Protocol;
-use crate::server::resource::Server;
 use crate::shared::replication::components::{DespawnTracker, Replicate};
 use crate::shared::time_manager::is_ready_to_send;
 use crate::utils::wrapping_id::wrapping_id;
@@ -465,18 +464,18 @@ fn clean_entity_despawns(
 
 #[cfg(test)]
 mod tests {
-    use bevy::utils::{Duration, HashMap};
-
     use bevy::ecs::system::RunSystemOnce;
     use bevy::prelude::Events;
+    use bevy::utils::{Duration, HashMap};
 
-    use super::*;
     use crate::prelude::client::*;
     use crate::prelude::*;
     use crate::shared::replication::components::ReplicationMode;
     use crate::tests::protocol::Replicate;
     use crate::tests::protocol::*;
     use crate::tests::stepper::{BevyStepper, Step};
+
+    use super::*;
 
     fn setup() -> BevyStepper {
         let frame_duration = Duration::from_millis(10);
