+++ lightyear/src/shared/events/mod.rs
@@ -1,0 +1,6 @@
+//! This module defines bevy [`Events`](bevy::prelude::Events) related to networking events
+pub(crate) mod connection;
+
+pub mod components;
+pub mod plugin;
+pub mod systems;
