--- lightyear/src/shared/systems/mod.rs
@@ -1,2 +1,0 @@
-//! Bevy [`System`](bevy::prelude::System) that are shared between the server and client
-pub mod events;
