+++ macros/src/message.rs
@@ -152,9 +152,7 @@ pub fn message_protocol_impl(
             use bevy::ecs::entity::{MapEntities, EntityMapper};
             use #shared_crate_name::_reexport::*;
             use #shared_crate_name::prelude::*;
-            use #shared_crate_name::shared::systems::events::push_message_events;
-            use #shared_crate_name::shared::events::{IterMessageEvent, MessageEvent};
-
+            use #shared_crate_name::shared::events::systems::push_message_events;
 
             #[derive(Serialize, Deserialize, Clone, PartialEq)]
             #extra_derives
